Finished home screen and moved screens to separate files
