+++ app/src/main/java/com/example/rpg_character_sheet/CharacterActionsScreen.kt
@@ -1,0 +1,11 @@
+package com.example.rpg_character_sheet
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.navigation.NavHostController
+
+
+@Composable
+fun CharacterActionsScreen(viewModel: CharacterViewModel, navController: NavHostController) {
+	val character = viewModel.getSelectedCharacter().collectAsState(initial = null).value
+}
